:recycle: Moved notifications into services
