+++ lib/build.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const task = require("./task");
-const notification = require("../lib/notification");
+const notification = require("../services/notification");
 
 /**
  * startBuild
