+++ lib/task.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const taskDetail = require("./taskDetail");
-const notification = require("./notification");
+const notification = require("../services/notification");
 const taskQueue = require("./taskQueue");
 const taskUpdate = require("./taskUpdate");
