+++ lib/taskUpdate.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const notification = require("../lib/notification");
+const notification = require("../services/notification");
 const task = require("./task");
 
 /**
